Add CSRF token to the method of the admin controllers and add refacto.

+++ src/Controller/Admin/FeatureAdminController.php
@@ -3,34 +3,31 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Feature;
-use App\Form\FeatureType;
 use App\Repository\FeatureRepository;
-use App\Service\IgdbApiService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/admin-{secret}/feature', name: 'app_admin_feature_')]
 #[IsGranted('ROLE_ADMIN')]
 class FeatureAdminController extends AbstractController
 {
-    private $entityManager;
-
-    public function __construct(EntityManagerInterface $entityManager)
+    public function __construct(private EntityManagerInterface $entityManager, private string $admin_secret)
     {
-        $this->entityManager = $entityManager;
     }
 
     // Méthode pour afficher toutes les fonctionnalités
     #[Route('/', name: 'show')]
     #[IsGranted('ROLE_ADMIN')]
-    public function showFeatures(string $secret, FeatureRepository $featureRepository): Response
+    public function showFeatures(FeatureRepository $featureRepository): Response
     {
         $expectedSecret = $this->getParameter('admin_secret');
-        if ($secret !== $expectedSecret) {
+        if ($this->admin_secret !== $expectedSecret) {
             throw $this->createAccessDeniedException('Page not found');
         }
 
@@ -45,10 +42,10 @@ class FeatureAdminController extends AbstractController
     // Méthode pour afficher les détails d'une fonctionnalité
     #[Route('/details/{id}', name: 'details')]
     #[IsGranted('ROLE_ADMIN')]
-    public function detailsFeature(string $secret, Feature $feature): Response
+    public function detailsFeature(Feature $feature): Response
     {
         $expectedSecret = $this->getParameter('admin_secret');
-        if ($secret !== $expectedSecret) {
+        if ($this->admin_secret !== $expectedSecret) {
             throw $this->createAccessDeniedException('Page not found');
         }
 
@@ -61,10 +58,10 @@ class FeatureAdminController extends AbstractController
     // Méthode pour supprimer une fonctionnalité
     #[Route('/delete/{id}', name: 'delete')]
     #[IsGranted('ROLE_ADMIN')]
-    public function deleteFeature(string $secret, Feature $feature): Response
+    public function deleteFeature(Feature $feature, CsrfTokenManagerInterface $csrfTokenManager, Request $request): Response
     {
         $expectedSecret = $this->getParameter('admin_secret');
-        if ($secret !== $expectedSecret) {
+        if ($this->admin_secret !== $expectedSecret) {
             throw $this->createAccessDeniedException('Page not found');
         }
 
@@ -72,9 +69,15 @@ class FeatureAdminController extends AbstractController
             throw $this->createNotFoundException('Feature not found');
         }
 
+        $token = new CsrfToken('delete_item', $request->request->get('_token'));
+
+        if (!$csrfTokenManager->isTokenValid($token)) {
+            throw $this->createAccessDeniedException('Token not valid');
+        }
+
         $this->entityManager->remove($feature);
         $this->entityManager->flush();
 
-        return $this->redirectToRoute('app_admin_feature_show', ['secret' => $secret]);
+        return $this->redirectToRoute('app_admin_feature_show', ['secret' => $this->admin_secret]);
     }
 }
